QDOC: move `ImportItem` from `AutoImportFix` to `RsQualifiedNamedElement.kt` as `QualifiedNamedItem`.
Purpose of the change it to generalize work with reexported items and use them in quick documentation
for correct link resolution
