+++ src/main/kotlin/org/rust/lang/core/psi/ext/RsQualifiedNamedElement.kt
@@ -6,9 +6,12 @@
 package org.rust.lang.core.psi.ext
 
 import com.intellij.openapi.diagnostic.Logger
+import com.intellij.openapi.project.Project
+import org.rust.cargo.project.workspace.CargoWorkspace
 import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.ext.RsQualifiedName.ChildItemType.*
 import org.rust.lang.core.psi.ext.RsQualifiedName.ParentItemType.*
+import org.rust.lang.core.stubs.index.RsReexportIndex
 import org.rust.lang.core.types.ty.TyPrimitive
 
 interface RsQualifiedNamedElement : RsNamedElement {
@@ -273,4 +276,106 @@ data class RsQualifiedName private constructor(
             }
         }
     }
+}
+
+sealed class QualifiedNamedItem(val item: RsQualifiedNamedElement) {
+
+    abstract val itemName: String?
+    abstract val isPublic: Boolean
+    abstract val superMods: List<RsMod>?
+    abstract val containingCargoTarget: CargoWorkspace.Target?
+
+    val parentCrateRelativePath: String? get() {
+        val path = superMods
+            ?.map { it.modName ?: return null }
+            ?.asReversed()
+            ?.drop(1)
+            ?.joinToString("::") ?: return null
+        return if (item is RsEnumVariant) item.parentEnum.name?.let { "$path::$it" } else path
+    }
+
+    val crateRelativePath: String? get() {
+        val name = itemName ?: return null
+        val parentPath = parentCrateRelativePath ?: return null
+        if (parentPath.isEmpty()) return name
+        return "$parentPath::$name"
+    }
+
+    class ExplicitItem(item: RsQualifiedNamedElement) : QualifiedNamedItem(item) {
+        override val itemName: String? get() = item.name
+        override val isPublic: Boolean get() = (item as? RsVisible)?.isPublic == true
+        override val superMods: List<RsMod>? get() = (if (item is RsMod) item.`super` else item.containingMod)?.superMods
+        override val containingCargoTarget: CargoWorkspace.Target? get() = item.containingCargoTarget
+    }
+
+    class ReexportedItem(
+        private val useSpeck: RsUseSpeck,
+        item: RsQualifiedNamedElement
+    ) : QualifiedNamedItem(item) {
+
+        override val itemName: String? get() = useSpeck.nameInScope
+        override val isPublic: Boolean get() = true
+        override val superMods: List<RsMod>? get() = useSpeck.containingMod.superMods
+        override val containingCargoTarget: CargoWorkspace.Target? get() = useSpeck.containingCargoTarget
+    }
+
+    class CompositeItem(
+        override val itemName: String?,
+        override val isPublic: Boolean,
+        private val reexportedModItem: QualifiedNamedItem,
+        private val explicitSuperMods: List<RsMod>,
+        item: RsQualifiedNamedElement
+    ) : QualifiedNamedItem(item) {
+
+        override val superMods: List<RsMod>? get() {
+            val mods = ArrayList(explicitSuperMods)
+            mods += reexportedModItem.superMods.orEmpty()
+            return mods
+        }
+        override val containingCargoTarget: CargoWorkspace.Target? get() = reexportedModItem.containingCargoTarget
+    }
+}
+
+/**
+ * Collect all possible imports using reexports of modules from original import item path
+ */
+fun QualifiedNamedItem.withModuleReexports(project: Project): List<QualifiedNamedItem> {
+    check(this is QualifiedNamedItem.ExplicitItem || this is QualifiedNamedItem.ReexportedItem) {
+        "`QualifiedNamedItem.withModuleReexports` should be called only for `QualifiedNamedItem.ExplicitItem` and `QualifiedNamedItem.ReexportedItem`"
+    }
+
+    // Contains already visited edges of module graph
+    // (useSpeck element <-> reexport edge of module graph).
+    // Only reexports can create cycles in module graph
+    // so it's enough to collect only such edges
+    val visited: MutableSet<RsUseSpeck> = HashSet()
+
+    fun QualifiedNamedItem.collectImportItems(): List<QualifiedNamedItem> {
+        val importItems = mutableListOf(this)
+        val superMods = superMods.orEmpty()
+        superMods
+            // only public items can be reexported
+            .filter { it.isPublic }
+            .forEachIndexed { index, ancestorMod ->
+                val modName = ancestorMod.modName ?: return@forEachIndexed
+                RsReexportIndex.findReexportsByName(project, modName)
+                    .mapNotNull {
+                        if (it in visited) return@mapNotNull null
+                        val reexportedMod = it.path?.reference?.resolve() as? RsMod
+                        if (reexportedMod != ancestorMod) return@mapNotNull null
+                        it to reexportedMod
+                    }
+                    .forEach { (useSpeck, reexportedMod) ->
+                        visited += useSpeck
+                        val items = QualifiedNamedItem.ReexportedItem(useSpeck, reexportedMod).collectImportItems()
+                        importItems += items.map {
+                            QualifiedNamedItem.CompositeItem(itemName, isPublic, it, superMods.subList(0, index + 1), item)
+                        }
+                        visited -= useSpeck
+                    }
+            }
+        return importItems
+    }
+
+    return collectImportItems()
 }
